refactor: theHeader
