+++ apps/frontend/src/composables/index.ts
@@ -2,4 +2,4 @@ export { useTaskOperationMessage } from './useTaskOperationMessage'
 export { useTaskRightContextMenu } from './useTaskRightContextMenu'
 export { useTaskSidebarDrag } from './useTaskSidebarDrag'
 export { useIsMac } from './misc'
-export { useGoto } from './goto'
+export * from './goto'
